agent/pyproject.toml
[project]
name = "lego-brickbase-agent"
version = "0.1.0"
description = "Agentic AI layer for LEGO BrickBase – natural language analytics over the Rebrickable LEGO dataset"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai[anthropic]>=0.2",
    "anthropic>=0.40",
    "pydantic>=2.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "streamlit>=1.45",
    "httpx>=0.28",
    "databricks-sql-connector>=3.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.9",
]
